refactor(server): extract route registration to boot/boot-routes.ts (#4243 step 5)
